assert-news: add `-u` command-line option
assert-news ignores uncommitted changes when invoked with `-l` option.
But `-l` is the option used in `pre-commit` hook tests which is used
before a commit a created. This means that news files will be
uncommitted at this stage. This is causing `pre-commit` hook failures
when invoked locally.

The `.pre-commit-config.yaml` has been updated to add `-u` option

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>